Implement cpu/gpu device selection
Pass a mode argument to the device initialization to choose the
execution mode. As with HOOMD, the default is auto, which selects GPUs
if they are available, otherwise it falls back to the CPU. The
difference here is that it is possible to create a build without the CPU
library, so there are error conditions to raise helpful exceptions if an
unavailable library were to be selected.

refs #3