Raise an AttributeError when modifying read-only materials
To solve the issue of attempting to modify local temporary copies of a
material, track if the material is a borrowed one and raise an attribute
error. This catches simple usage like this:

   tris.material.solid = 4 # raises error

Though, it also catches this:

   m = tris.material
   m.solid = 4 # raises error

Materials can still be read from one geometry and written to another

   squares.material = tris.material

they just can't be modified in between.