Prototype outlines
Compute distance to the nearest edge in the new prism geomtry. Use this
distance with a hard-coded edge width and color (black) in TracerWhitted
on the CPU.

The algorithm to determine the distance to the edge is general and will
work with any covnex polyhedra. It works by looking at the intersection
of the hit plane with all of the other planes in the shape. The distance
to the nearest edge is the smallest distance from the hit point to one
of these intersection lines.

I have not done any performance benchmarks of this method. Once we
establish a testing methodology, we can revisit this code and benchmark
it with edges on and off - and also profile it and solve obvious
hotspots.

The result of the current implementation amounts to hidden line removal.
We could also add a trivial extension for full wireframe
representations. Simply return with no interesection if the ray failes
to hit within the edge region (when the user activates this mode, of
course). This would make the polyhedron shape be a simple shell of
edges.