Autoset ENABLE_library options ON or OFF
Based on whether a library is found on first configure, automatically
set the corresponding ENABLE_library option on or off. This requires
less user intervention to enable desired libraries - under the
assumption that if the library is available to CMake, the user must want
to use it.

Users can always override with -DENABLE_library=OFF.

refs #39