Implement gamma-correct output on the gPU
Add a sRGB function to ColorMath that converts RGBA<float> in linear
space to RGBA<unsigned char> in sRGB space. Use this method in the
direct tracers to map output colors. Also implement two output buffers
in tracer, one in linear space and one in sRGB space. The linear space
one will be important later when averaging frames together in the
sampling tracer.

refs #9