Limit the number of threads used
Allow the user to pass in a thread limit to Device. Using
tbb::task_arena, limit the number of threads that embree uses in
rtcCommit and that fresnel uses to render on the CPU.

The limit API can be made more general and apply to the GPU in a later
update.