Implement full set of unit tests
These unit tests are automated versions of the example notebooks. They
render a low resolution image and compare it against a reference image.

For a given test (e.g. test_material), run ``python test_material.py``
to generate the images in ``output/`` for inspection. Modify
the file's __main__ to use a specific device (i.e. to debug why a
specific device fails the test). When a reference image is validated,
copy it to reference/ which the tests will use as a comparison.

refs #13