Determine module availability at compile time
This avoids problems where (e.g.) _cpu is in the fresnel directory, but
not compiled to the current up to date code. With the new determination,
baked in compile time info is used to determine whether or not _cpu
should be imported.

Also add a user-facing API to query what modes are available.