Device can now limit how many GPUs it uses
This doesn't fully allow GPU selection, but that can trivially be done
with environment variables. This update enables the most used option of
controlling how many GPUs, which is useful in node with N GPUs of the
same type. It also makes the API consistent between CPU and GPU devices.

fixes #36