Use numpy.unique for face merging
When I was using the dot products to compare if two normal vectors were
equal, I assumed that normal vectors were returned from qhull, but that
is not the case. This commit changes the code to use numpy's unique
feature; this change makes the code much cleaner, but we may have to do
some sensivity tests to see how similar things are before they're
considered the same.

I found this mistake by writing unit tests, specifically for the
dodecahedron.